backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]>=0.34.0
google-adk==0.1.0
python-dotenv==1.0.1
pydantic==2.10.4
supabase>=2.8.0
google-genai>=1.20.0
numpy>=1.24.0
scikit-learn>=1.3.0